Change set-font to error if font cannot be fonud
I defined a corresponding font-not-found condition.

I also changed set-font to repeat itself less. Instead of checking if
font is a list and having slightly different logic for a list and an
atom, set-font now calls the helper alexandria:ensure-list and retains
only the logic for the cases wher efont is a list.